move UseLocaleExtension to the proper package
it was declared in "integration" package, but used in unit-tests.
