+++ src/test/java/com/codeborne/selenide/UseLocaleExtension.java
@@ -1,5 +1,4 @@
-package integration;
-
+package com.codeborne.selenide;
 
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.api.extension.InvocationInterceptor;
