+++ src/test/java/com/codeborne/selenide/impl/DurationFormatTest.java
@@ -1,6 +1,6 @@
 package com.codeborne.selenide.impl;
 
-import integration.UseLocaleExtension;
+import com.codeborne.selenide.UseLocaleExtension;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
