+++ src/test/java/com/codeborne/selenide/logevents/SelenideLoggerTest.java
@@ -1,6 +1,6 @@
 package com.codeborne.selenide.logevents;
 
-import integration.UseLocaleExtension;
+import com.codeborne.selenide.UseLocaleExtension;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
